model/requirements.txt
aiofiles
cryptography
fastapi
google-generativeai
groq
gTTS
lark
langchain
langchain_chroma
langchain_community
langchain_core
langchain_groq
langchain_huggingface
langchain_openai
mysql-connector-python
opencv_python
openpyxl
Pillow
pydantic
pydub
PyMuPDF
pymysql
pytesseract
python-docx
python-dotenv
python-multipart
python-pptx
Requests
twilio
unstructured
uvicorn
passlib

langchain_experimental
neo4j

httpx
pytest
pytest-asyncio